PLAT-51382: Syncing DOM structure and logic between JS version and Native version lists
Enact-DCO-1.0-Signed-off-by: Seungcheon Baek (sc.baek@lge.com)
