+++ packages/moonstone/VirtualList/VirtualListBase.js
@@ -59,7 +59,7 @@ const VirtualListBaseFactory = (type) => {
 	const UiBase = (type === JS) ? UiVirtualListBase : UiVirtualListBaseNative;
 
 	return class VirtualListCore extends Component {
-		static displayName = 'VirtualListBase'
+		/* No displayName here. We set displayName to returned components of this factory function. */
 
 		static propTypes = /** @lends moonstone/VirtualList.VirtualList.prototype */ {
 			/**
@@ -130,9 +130,9 @@ const VirtualListBaseFactory = (type) => {
 		}
 
 		componentDidMount () {
-			if (type === JS) {
-				const containerNode = this.uiRef.containerRef;
+			const containerNode = this.uiRef.containerRef;
 
+			if (type === JS) {
 				// prevent native scrolling by Spotlight
 				this.preventScroll = () => {
 					containerNode.scrollTop = 0;
@@ -141,14 +141,11 @@ const VirtualListBaseFactory = (type) => {
 
 				if (containerNode && containerNode.addEventListener) {
 					containerNode.addEventListener('scroll', this.preventScroll);
-					containerNode.addEventListener('keydown', this.onKeyDown);
 				}
-			} else {
-				const contentNode = this.uiRef.contentRef;
+			}
 
-				if (contentNode && contentNode.addEventListener) {
-					contentNode.addEventListener('keydown', this.onKeyDown);
-				}
+			if (containerNode && containerNode.addEventListener) {
+				containerNode.addEventListener('keydown', this.onKeyDown);
 			}
 		}
 
@@ -157,20 +154,18 @@ const VirtualListBaseFactory = (type) => {
 		}
 
 		componentWillUnmount () {
-			if (type === JS) {
-				const containerNode = this.uiRef.containerRef;
+			const containerNode = this.uiRef.containerRef;
 
+			if (type === JS) {
 				// remove a function for preventing native scrolling by Spotlight
 				if (containerNode && containerNode.removeEventListener) {
 					containerNode.removeEventListener('scroll', this.preventScroll);
 					containerNode.removeEventListener('keydown', this.onKeyDown);
 				}
-			} else {
-				const contentNode = this.uiRef.contentRef;
+			}
 
-				if (contentNode && contentNode.removeEventListener) {
-					contentNode.removeEventListener('keydown', this.onKeyDown);
-				}
+			if (containerNode && containerNode.removeEventListener) {
+				containerNode.removeEventListener('keydown', this.onKeyDown);
 			}
 
 			this.setContainerDisabled(false);
@@ -184,7 +179,7 @@ const VirtualListBaseFactory = (type) => {
 		restoreLastFocused = false
 
 		setContainerDisabled = (bool) => {
-			const containerNode = (type === JS) ? this.uiRef.containerRef : this.uiRef.contentRef;
+			const containerNode = this.uiRef.containerRef;
 
 			if (containerNode) {
 				containerNode.setAttribute(dataContainerDisabledAttribute, bool);
@@ -330,12 +325,6 @@ const VirtualListBaseFactory = (type) => {
 					isRtl = this.props.rtl,
 					isForward = (direction === 'down' || isRtl && direction === 'left' || !isRtl && direction === 'right');
 
-				if (type === JS) {
-					// To prevent item positioning issue, make all items to be rendered.
-					this.uiRef.updateFrom = null;
-					this.uiRef.updateTo = null;
-				}
-
 				if (firstVisibleIndex <= indexToScroll && indexToScroll <= lastVisibleIndex) {
 					const node = this.uiRef.containerRef.querySelector(`[data-index='${indexToScroll}'].spottable`);
 
@@ -479,9 +468,7 @@ const VirtualListBaseFactory = (type) => {
 
 			this.isScrolledBy5way = false;
 			if (getDirection(keyCode)) {
-				if (type === Native) {
-					ev.preventDefault();
-				}
+				ev.preventDefault();
 				this.setSpotlightContainerRestrict(keyCode, target);
 				this.isScrolledBy5way = this.jumpToSpottableItem(keyCode, target);
 			}
@@ -511,9 +498,6 @@ const VirtualListBaseFactory = (type) => {
 
 			if (Spotlight.isPaused()) {
 				Spotlight.resume();
-				if (type === JS) {
-					this.forceUpdate();
-				}
 			}
 			this.focusOnNode(item);
 			this.nodeIndexToBeFocused = null;
@@ -523,7 +507,7 @@ const VirtualListBaseFactory = (type) => {
 			if (ref) {
 				if (type === JS) {
 					this.focusOnItem(index);
-				} else if (type === Native) {
+				} else {
 					// If focusing the item of VirtuallistNative, `onFocus` in Scrollable will be called.
 					// Then VirtualListNative tries to scroll again differently from VirtualList.
 					// So we would like to skip `focus` handling when focusing the item as a workaround.
@@ -624,9 +608,6 @@ const VirtualListBaseFactory = (type) => {
 					if (node) {
 						node.blur();
 					}
-				}
-				if (type === JS) {
-					this.nodeIndexToBeFocused = null;
 				}
 				this.lastFocusedIndex = focusedIndex;
 
@@ -734,6 +715,7 @@ const VirtualListBaseFactory = (type) => {
  * @private
  */
 const VirtualListBase = VirtualListBaseFactory(JS);
+VirtualListBase.displayName = 'VirtualListBase';
 
 /**
  * A Moonstone-styled base component for [VirtualListNative]{@link moonstone/VirtualList.VirtualListNative} and
@@ -746,6 +728,7 @@ const VirtualListBase = VirtualListBaseFactory(JS);
  * @private
  */
 const VirtualListBaseNative = VirtualListBaseFactory(Native);
+VirtualListBaseNative.displayName = 'VirtualListBaseNative';
 
 const ScrollableVirtualList = ({role, ...rest}) => ( // eslint-disable-line react/jsx-no-bind
 	<Scrollable
