--- packages/ui/VirtualList/ListItem.less
@@ -1,6 +1,0 @@
-.listItem {
-	overflow: hidden;
-	width: 100%;
-
-	will-change: transform;
-}
