+++ packages/ui/VirtualList/VirtualList.less
@@ -2,6 +2,7 @@
 	position: relative;
 	height: 100%;
 	width: 100%;
+	overflow: hidden;
 
 	&.vertical {
 		overflow-x: hidden;
@@ -11,6 +12,13 @@
 	&.horizontal {
 		overflow-x: scroll;
 		overflow-y: hidden;
+	}
+
+	.listItem {
+		overflow: hidden;
+		width: 100%;
+
+		will-change: transform;
 	}
 }
